move test case from tests/integrations to tests/failpoints
